Crate azure_iot_sdk

Wrapper around azure iot-c-sdk-ref.

A reference implementation can be found here.//!

Provides an abstraction over Microsoft’s iot-c-sdk in order to develop device- and module twin client applications. All API’s exposed by this crate base on the following low level function interfaces:

module client

device client

The following use cases can be realized by using this crate:

connect as module or device client to azure iothub

either use directly a connection string

use iot-identity-service (if installed on device) to create the client and build the connection string. Note: This feature is currently not supported for all combinations of identity type and authentication mechanism.

module twin or device twin twin based communication with iothub:

read/write tags

receive desired properties

send reported properties

direct methods

Device2Cloud (D2C) messages

Cloud2Device (C2D) messages

Modules

client

iothub client Let’s you create an instance of Self::IotHubClient with integrated Self::EventHandler.