Trait azure_iot_sdk::client::EventHandler

pub trait EventHandler {
    fn handle_connection_status(&self, auth_status: AuthenticationStatus) { ... }

    fn handle_c2d_message(&self, message: IotMessage) -> Result<(), IotError> { ... }

    fn get_c2d_message_property_keys(&self) -> Vec<&'static str> { ... }

    fn handle_twin_desired(
        &self, 
        state: TwinUpdateState, 
        desired: Value
    ) -> Result<(), IotError> { ... }

    fn get_direct_methods(&self) -> Option<&DirectMethodMap> { ... }
}

Trait to be implemented by client implementations in order to handle iothub events

struct MyEventHandler {}

impl EventHandler for MyEventHandler {
  fn handle_connection_status(&self, auth_status: AuthenticationStatus) {
      debug!("{:?}", auth_status)
  }

  fn handle_c2d_message(&self, message: IotMessage) -> Result<(), IotError> {
      debug!("{:?}", message);
      Ok(())
  }

  fn get_c2d_message_property_keys(&self) -> Vec<&'static str> {
      debug!("get message properties called");
      vec![]
  }

  fn handle_twin_desired(
      &self,
      state: TwinUpdateState,
      desired: serde_json::Value,
  ) -> Result<(), IotError> {
      debug!("{:?}, {:?}", state, desired);
      Ok(())
  }

  fn get_direct_methods(&self) -> Option<&DirectMethodMap> {
      debug!("get direct methods called");
      None
  }
}

Provided methods

fn handle_connection_status(&self, auth_status: AuthenticationStatus)

gets called as a result of callig IotHubClient::from_identity_service() or IotHubClient::from_connection_string()

fn handle_c2d_message(&self, message: IotMessage) -> Result<(), IotError>

gets called if a message from iothub is received

fn get_c2d_message_property_keys(&self) -> Vec<&'static str>

gets called in order to return mqtt property keys to extracted from an iothub message

fn handle_twin_desired(
&self,
state: TwinUpdateState,
desired: Value
) -> Result<(), IotError>

gets called if new desired properties were sent by iothub

fn get_direct_methods(&self) -> Option<&DirectMethodMap>

gets called if a direct method was called by iothub. client implementation must return a map with implemented direct methods.